+++ CHANGELOG.md
@@ -23,6 +23,17 @@ Each section organizes entries into the following subsections:
 [Unreleased]
 ------------
 
+### Dynamicbeat
+
+#### Changed
+
+- Dynamicbeat is now a standalone program that doesn't use libbeat (#302)
+
+#### Removed
+
+- Remove `update_period` setting, update configurations after starting each round (#302)
+- Drop Logstash from architecture (#302)
+
 [0.7.0] - 2020-02-21
 --------------------